pyproject.toml
[project]
name = "elastic-incident-commander"
version = "0.1.0"
description = "DevOps Incident Commander — Multi-agent incident response system built with Elastic Agent Builder"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
dependencies = [
    "elasticsearch>=8.17.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[project.scripts]
incident-commander = "incident_commander.cli:app"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
